requirements.txt
boto3
requests
PyYAML
cfn-flip